+++ frontend/src/App.jsx
@@ -1,4 +1,4 @@
-import LoginPage from "./pages/LoginPage";
+import LoginPage from "./pages/Login/LoginPage";
 
 function App() {
   return (
