+++ frontend/src/components/Login/Login.jsx
@@ -7,11 +7,15 @@ function Login() {
 
   const handleToggle = () => {
     setIsRegister(!isRegister);
-    setForgotPassword(false);
+    setForgotPassword(false); // Reset forgot password state when toggling between login and register
   };
 
   const handleForgotPasswordClick = () => {
     setForgotPassword(true);
+  };
+
+  const handleGoBack = () => {
+    setForgotPassword(false); // Return to the previous form
   };
 
   return (
@@ -78,6 +82,14 @@ function Login() {
           <h2 className="title">Enter your email for password reset</h2>
           <input type="email" placeholder="Email" className="input" />
           <button className="resetButton">Reset</button>
+          {/* New Go Back Button */}
+          <button className="goBackButton" onClick={handleGoBack}>
+            <img
+              src="https://cdn-icons-png.flaticon.com/512/93/93634.png"
+              alt="Back"
+              className="backIcon"
+            />
+          </button>
         </div>
       )}
     </div>
