+++ frontend/src/pages/Login/LoginPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Background from "../components/Background";
-import Login from "../components/Login";
+import Background from "../../components/Login/Background";
+import Login from "../../components/Login/Login";
 import "./LoginPage.css";
 
 const LoginPage = () => {
